split normalizer tests
