+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/compiler/normalizer/BoolMatcherSpec.scala
@@ -1,0 +1,20 @@
+package coop.rchain.rholang.interpreter.compiler.normalizer
+
+import coop.rchain.rholang.syntax.rholang_mercury.Absyn._
+
+import org.scalatest._
+
+import coop.rchain.models.Expr.ExprInstance._
+
+class BoolMatcherSpec extends FlatSpec with Matchers {
+  "BoolTrue" should "Compile as GBool(true)" in {
+    val btrue = new BoolTrue()
+
+    BoolNormalizeMatcher.normalizeMatch(btrue) should be(GBool(true))
+  }
+  "BoolFalse" should "Compile as GBool(false)" in {
+    val bfalse = new BoolFalse()
+
+    BoolNormalizeMatcher.normalizeMatch(bfalse) should be(GBool(false))
+  }
+}
